Made fps changable from the command line.
