+++ src/cray/crconsts.h
@@ -5,14 +5,14 @@
 
 static const int CRAY_SUCCESS = 0;
 static const int CRAY_FAILURE = 1;
-static const int CRAY_FPS = 60;
-static const int CRAY_SCREEN_WIDTH = 640;
-static const int CRAY_SCREEN_HEIGHT = 480;
-static const int CRAY_SCREEN_BPP = 32;
+static const unsigned int CRAY_DEFAULT_FPS = 60;
+static const unsigned int CRAY_DEFAULT_WIDTH = 640;
+static const unsigned int CRAY_DEFAULT_HEIGHT = 480;
+static const unsigned int CRAY_DEFAULT_BPP = 32;
 static const double PLAYER_ARROW_SIZE = 25.0;
 static const double PLAYER_BASE_SIZE = 6.0;
 static const double TRANS_SPEED = 2.0;
 static const double ROT_SPEED = 2.0;
-static const int WALL_HEIGHT = 4000.0;
+static const unsigned int WALL_HEIGHT = 4000.0;
 
 #endif // !_CR_CONSTS_H_
