+++ src/raycastdemo/raysettings.h
@@ -13,6 +13,12 @@ typedef struct RaycastSettings {
 	bool printDebugInfo;
 	RenderMode renderMode;
 	ScreenFormat screenFormat;
+	unsigned int targetFps;
 } RaycastSettings;
+
+RaycastSettings ParseCommandLine(
+	int argc,
+	char* argv[]
+);
 
 #endif // !_CRAY_SETTINGS_H_
